fix(ui): keep sub-pixel precision in the pointer position
The backend truncated gio's float pointer position to int before handing it to
the engine, a leftover of ebiten's integer CursorPosition. Element bounds are
fractional — layout under non-integer display scaling produces them everywhere —
so an integer cursor lands on the wrong side of an edge by up to a pixel. Hover
and clicks near a border trigger late, and nudging the mouse along an edge flips
the truncated value back and forth across the fractional bound, which reads as
the hover highlight flickering on and off.

Carry the position through as float32, the same rule the scroll amount already
follows: gio's units and precision are passed through, not re-derived.

Consumers move to float too — the hover cache and the multi-click distance
check, whose 4px threshold now compares actual distance rather than truncated
pixels.

+++ pkg/ui/input.go
@@ -43,7 +43,10 @@ const (
 )
 
 type inputSource interface {
-	cursor() (int, int)             // 光标位置（物理像素）
+	// cursor 返回光标位置：物理像素，保留亚像素精度。
+	// 不可取整 —— 元素 bounds 是小数（非整数缩放下尤其如此），整数光标在边界处会差到 1px，
+	// 表现为「贴着边框时 hover 时有时无」。
+	cursor() (float32, float32)
 	wheel() (float32, float32)      // 本帧滚轮增量（x,y），单位为物理像素，正值向右/向下
 	mousePressed(mouseBtn) bool     // 按钮当前是否按住
 	mouseJustPressed(mouseBtn) bool // 按钮本帧是否刚按下
